web/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
cfgv==3.4.0
distlib==0.3.8
Django==4.2.8
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
filelock==3.13.1
flake8==6.1.0
identify==2.5.33
importlib-metadata==7.0.0
isort==5.13.2
Markdown==3.5.1
mccabe==0.7.0
nodeenv==1.8.0
platformdirs==4.1.0
pre-commit==3.6.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
virtualenv==20.25.0
zipp==3.17.0
